Adding checkouts stub, setup now skips form
