+++ src/main/java/euromocks/Constants.java
@@ -4,6 +4,9 @@ package euromocks;
  * Created by sriramangajala on 26/01/16.
  */
 public class Constants {
-    public static String URL = "stubs-services.trainz.io";
-    public static int PORT = 8080;
+//    public static String URL = "stubs-services.trainz.io";
+//    public static int PORT = 8080;
+
+    public static String URL = "127.0.0.1";
+    public static int PORT = 9090;
 }
