+++ src/main/java/euromocks/Enovation_payment.java
@@ -94,7 +94,7 @@ public class Enovation_payment {
         try {
             FileReader fileReader = new FileReader("src/main/resources/euromocks/enovation/querypayment_200.json");
             JSONObject json = (JSONObject) parser.parse(fileReader);
-            stubFor(get(urlPathMatching("/enov/payments/query"))
+            stubFor(get(urlPathMatching("/enov/payments/query/*"))
 //                    .withRequestBody(containing("merchantReference"))
 //                    .withRequestBody(containing("\"pageSetId\":\"1\""))
 //                    .withRequestBody(containing("returnUrl"))
@@ -104,6 +104,24 @@ public class Enovation_payment {
                                     .withHeader("Access-Control-Allow-Headers", "accept, cid")
                                     .withHeader("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
                     ).atPriority(34));
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+
+    void checkout_200() throws IOException {
+        JSONParser parser = new JSONParser();
+        try {
+            FileReader fileReader = new FileReader("src/main/resources/euromocks/enovation/checkout_200.json");
+            JSONObject json = (JSONObject) parser.parse(fileReader);
+            stubFor(post(urlPathMatching("/enov/checkouts?.*"))
+                    .willReturn(aResponse().withStatus(200).withHeader("Content-Type", "application/json").withBody(json.toJSONString())
+                            .withHeader("Access-Control-Allow-Origin", "*")
+                            .withHeader("Access-Control-Allow-Credentials", "true")
+                            .withHeader("Access-Control-Allow-Headers", "accept, cid")
+                            .withHeader("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
+                    ).atPriority(32));
         } catch (Exception e) {
             e.printStackTrace();
         }
