+++ src/main/resources/euromocks/enovation/checkout_200.json
@@ -1,0 +1,36 @@
+{
+  "transactionId" : "11431",
+  "contactFirstName" : "Daniel",
+  "contactLastName" : "Choudhury",
+  "outbound" : {
+    "destinationCode" : "8727100",
+    "originCode" : "7015400",
+    "departureDate" : "2016-05-11",
+    "timeSlot" : "PM",
+    "journeyPrice" : 25
+  },
+  "inbound" : {
+    "destinationCode" : "7015400",
+    "originCode" : "8727100",
+    "departureDate" : "2016-05-14",
+    "timeSlot" : "PM",
+    "journeyPrice" : 25
+  },
+  "passengers" : [ {
+    "firstName" : "Mobileuser",
+    "lastName" : "Eurostar",
+    "passengerId" : "1",
+    "title" : "Mrs"
+  }, {
+    "firstName" : "sdfjng",
+    "lastName" : "asgd",
+    "passengerId" : "2",
+    "title" : "Mrs"
+  } ],
+  "playerId" : "2",
+  "gameCode" : "GAME1",
+  "globalPrice" : 100,
+  "referenceCode" : "EN01-5XI5-KWDS-8HS9",
+  "contactEmailAddress" : "mobile.eurostar@gmail.com",
+  "confirmed" : true
+}
